+++ backend/packages/workspaces/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/workspaces
 
+## 0.7.43
+
+### Patch Changes
+
+- Updated dependencies [17adf4c]
+  - @cat-factory/contracts@0.33.0
+  - @cat-factory/kernel@0.36.0
+
 ## 0.7.42
 
 ### Patch Changes